add alert meal delete

+++ src/screens/MealDetails/index.tsx
@@ -17,6 +17,7 @@ import { HeaderScreen } from "@components/HeaderScreen";
 import { Button } from "@components/Button";
 import { Alert } from "react-native";
 import { deleteMeal } from "@storage/meals/deleteMeal";
+import { ModalDelete } from "@components/ModalDelete";
 
 export type MealDataProps = {
   meal: {
@@ -44,9 +45,17 @@ export function MealDetails() {
 
   async function handleDeleteMeal() {
     try {
-      console.log("deleteMeal");
-      await deleteMeal(day, data.hour);
-      navigation.navigate("home");
+      Alert.alert("", "Deseja realmente excluir o registro da refeição?", [
+        { text: "Não", style: "cancel" },
+        {
+          text: "Sim",
+          style: "default",
+          onPress: async () => {
+            await deleteMeal(day, data.hour);
+            navigation.navigate("home");
+          },
+        },
+      ]);
     } catch (error) {
       console.error(error);
       Alert.alert("Excluir refeição", "Não foi possível excluir refeição");
